friendlyface/recognition/detection.py
"""Face detection, landmark alignment, and quality scoring (US-045).

Provides a unified ``FaceDetector`` that uses MediaPipe when available
and falls back to a Haar-cascade + heuristic detector otherwise.
Detected faces are aligned to a canonical 112x112 crop using 5-point
landmarks (eyes, nose, mouth corners).
"""

from __future__ import annotations

import logging
from dataclasses import dataclass, field

import numpy as np
from PIL import Image

logger = logging.getLogger("friendlyface.recognition.detection")

# Target output size after alignment
ALIGNED_SIZE = (112, 112)

try:
    import mediapipe as mp  # type: ignore[import-untyped]

    _HAS_MEDIAPIPE = True
except ImportError:
    _HAS_MEDIAPIPE = False


@dataclass
class FaceLandmarks:
    """5-point facial landmarks (pixel coordinates)."""

    left_eye: tuple[float, float]
    right_eye: tuple[float, float]
    nose_tip: tuple[float, float]
    mouth_left: tuple[float, float]
    mouth_right: tuple[float, float]


@dataclass
class DetectedFace:
    """A single detected face with bounding box, landmarks, and quality."""

    bbox: tuple[int, int, int, int]  # (x, y, w, h)
    landmarks: FaceLandmarks | None
    quality_score: float  # 0.0 – 1.0
    confidence: float  # detector confidence
    aligned: np.ndarray | None = field(default=None, repr=False)


def _compute_quality(image_array: np.ndarray) -> float:
    """Estimate face image quality from sharpness and brightness.

    Returns a score in [0, 1] where 1 = ideal quality.
    """
    gray = image_array
    if gray.ndim == 3:
        # Simple luminance conversion
        gray = np.mean(image_array, axis=2)

    # Sharpness via Laplacian variance (approximated with finite diffs)
    laplacian = (
        gray[:-2, 1:-1] + gray[2:, 1:-1] + gray[1:-1, :-2] + gray[1:-1, 2:] - 4 * gray[1:-1, 1:-1]
    )
    sharpness = float(np.var(laplacian))
    # Normalize: typical sharp face has laplacian var > 500
    sharpness_score = min(sharpness / 500.0, 1.0)

    # Brightness: ideal mean around 127
    mean_brightness = float(np.mean(gray))
    brightness_score = 1.0 - abs(mean_brightness - 127.0) / 127.0

    # Contrast: std dev of pixel values
    contrast = float(np.std(gray))
    contrast_score = min(contrast / 64.0, 1.0)

    return round(0.5 * sharpness_score + 0.3 * brightness_score + 0.2 * contrast_score, 4)


def _align_face(
    image: np.ndarray,
    landmarks: FaceLandmarks,
    output_size: tuple[int, int] = ALIGNED_SIZE,
) -> np.ndarray:
    """Align a face crop using eye-center rotation + scale.

    Rotates the image so the eye line is horizontal and crops to
    ``output_size``.
    """
    le = np.array(landmarks.left_eye)
    re = np.array(landmarks.right_eye)

    eye_center = (le + re) / 2.0
    dy = re[1] - le[1]
    dx = re[0] - le[0]
    angle = float(np.degrees(np.arctan2(dy, dx)))

    pil_img = Image.fromarray(image.astype(np.uint8))
    rotated = pil_img.rotate(-angle, center=(float(eye_center[0]), float(eye_center[1])))
    # Crop centered on eye midpoint
    cx, cy = float(eye_center[0]), float(eye_center[1])
    half_w, half_h = output_size[0] / 2, output_size[1] / 2
    box = (
        max(int(cx - half_w), 0),
        max(int(cy - half_h * 0.7), 0),  # shift up slightly (forehead)
        min(int(cx + half_w), rotated.width),
        min(int(cy + half_h * 1.3), rotated.height),
    )
    cropped = rotated.crop(box).resize(output_size, Image.Resampling.LANCZOS)
    return np.asarray(cropped)


class FaceDetector:
    """Face detection with optional MediaPipe backend.

    Falls back to a simple intensity-based detector when MediaPipe
    is not installed — useful for testing and lightweight deployments.
    """

    def __init__(self, min_confidence: float = 0.5) -> None:
        self.min_confidence = min_confidence
        self._mp_detector = None

        if _HAS_MEDIAPIPE:
            self._init_mediapipe()

    def _init_mediapipe(self) -> None:
        """Initialize MediaPipe face detection."""
        mp_face = mp.solutions.face_detection  # type: ignore[attr-defined]
        self._mp_detector = mp_face.FaceDetection(
            model_selection=1,
            min_detection_confidence=self.min_confidence,
        )
        logger.info("MediaPipe face detector initialized")

    @property
    def backend(self) -> str:
        """Return the active detection backend name."""
        return "mediapipe" if self._mp_detector is not None else "heuristic"

    def detect(self, image: np.ndarray) -> list[DetectedFace]:
        """Detect faces in an image array (H, W, 3) RGB.

        Returns a list of ``DetectedFace`` instances sorted by
        confidence descending.
        """
        if self._mp_detector is not None:
            return self._detect_mediapipe(image)
        return self._detect_heuristic(image)

    def _detect_mediapipe(self, image: np.ndarray) -> list[DetectedFace]:
        """Detect faces using MediaPipe."""
        results = self._mp_detector.process(image)  # type: ignore[union-attr]
        if not results.detections:
            return []

        h, w = image.shape[:2]
        faces: list[DetectedFace] = []

        for det in results.detections:
            bb = det.location_data.relative_bounding_box
            x = int(bb.xmin * w)
            y = int(bb.ymin * h)
            bw = int(bb.width * w)
            bh = int(bb.height * h)

            kps = det.location_data.relative_keypoints
            landmarks = FaceLandmarks(
                left_eye=(kps[0].x * w, kps[0].y * h),
                right_eye=(kps[1].x * w, kps[1].y * h),
                nose_tip=(kps[2].x * w, kps[2].y * h),
                mouth_left=(kps[3].x * w, kps[3].y * h),
                mouth_right=(kps[4].x * w, kps[4].y * h),
            )

            # Crop face region for quality assessment
            face_crop = image[max(y, 0) : y + bh, max(x, 0) : x + bw]
            quality = _compute_quality(face_crop) if face_crop.size > 0 else 0.0
            conf = float(det.score[0])

            aligned = _align_face(image, landmarks) if conf >= self.min_confidence else None

            faces.append(
                DetectedFace(
                    bbox=(x, y, bw, bh),
                    landmarks=landmarks,
                    quality_score=quality,
                    confidence=conf,
                    aligned=aligned,
                )
            )

        return sorted(faces, key=lambda f: f.confidence, reverse=True)

    def _detect_heuristic(self, image: np.ndarray) -> list[DetectedFace]:
        """Simple center-crop heuristic detector (no external deps).

        Assumes a single face centered in the image. Useful for testing
        and environments without MediaPipe.
        """
        h, w = image.shape[:2]
        # Assume face occupies central 60% of image
        margin_x = int(w * 0.2)
        margin_y = int(h * 0.15)
        bw = w - 2 * margin_x
        bh = h - 2 * margin_y

        # Synthetic landmarks based on face proportions
        cx, cy = w / 2, h / 2
        landmarks = FaceLandmarks(
            left_eye=(cx - bw * 0.15, cy - bh * 0.1),
            right_eye=(cx + bw * 0.15, cy - bh * 0.1),
            nose_tip=(cx, cy + bh * 0.05),
            mouth_left=(cx - bw * 0.1, cy + bh * 0.2),
            mouth_right=(cx + bw * 0.1, cy + bh * 0.2),
        )

        face_crop = image[margin_y : margin_y + bh, margin_x : margin_x + bw]
        quality = _compute_quality(face_crop) if face_crop.size > 0 else 0.0

        aligned = _align_face(image, landmarks)

        return [
            DetectedFace(
                bbox=(margin_x, margin_y, bw, bh),
                landmarks=landmarks,
                quality_score=quality,
                confidence=0.95,
                aligned=aligned,
            )
        ]
