pyproject.toml
[project]
name = "friendlyface"
version = "0.1.0"
description = "Forensic-Friendly AI Facial Recognition Platform — Layer 3: Blockchain Forensic Layer"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Dicoangelo" }]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "aiosqlite>=0.20.0",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "Pillow>=10.0.0",
    "python-multipart>=0.0.9",
    "lime>=0.2.0",
    "supabase>=2.0.0",
    "pynacl>=1.5.0",
    "python-json-logger>=3.0.0",
    "pydantic-settings>=2.0.0",
    "slowapi>=0.1.9",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "pyjwt[crypto]>=2.8.0",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
ml = [
    "mediapipe>=0.10.0",
    "onnxruntime>=1.17.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "bandit>=1.8.0",
]

[project.scripts]
app = "friendlyface.api.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["friendlyface"]
omit = ["friendlyface/__main__.py"]

[tool.coverage.report]
fail_under = 80
show_missing = true
skip_empty = true

[tool.ruff]
target-version = "py311"
line-length = 100
